perf: 暂存优化 vault table
